requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.3
